@@ -7,47 +7,45 @@ import (
 	"time"
 )
 
-func TestConcurrency(t *testing.T) {
-	// Create a thread pool that can allocate up to 10 threads
-	pool := New(10, func(params ...interface{}) {
-		n := params[0].(int)
-		w := params[1].(string)
-		time.Sleep(time.Second)
-		fmt.Println(w, n)
-	})
-	defer pool.Close()
-
-	for i := 0; i < 100; i++ {
-		fmt.Println("a:", i)
-		pool.Process(i, "hello")
-	}
-	pool.Wait()
-
-	for i := 0; i < 20; i++ {
-		fmt.Println("b:", i)
-		pool.Process(i, "world")
-	}
-
-	// Multiple threads will be notified at the same time
-	go func() {
-		pool.Wait()
-		fmt.Println("wait on go thread ")
-	}()
-
-	pool.Wait()
-	fmt.Println("wait on main thread")
-}
-
 func TestSingleTask(t *testing.T) {
-	pool := New(10, func(params ...interface{}) {
+	uptime := time.Now()
+	pool := New(10, func(params ...any) {
 		task := params[0].(int)
 		time.Sleep(time.Second)
 		fmt.Println("run", task)
 	})
 	defer pool.Close()
-	for i := 0; i < 10000; i++ {
+	for i := 0; i < 10; i++ {
+		pool.Process(i)
+	}
+
+	go func() {
+		pool.Wait()
+		log.Println("goroutine")
+	}()
+
+	pool.Wait()
+	pool.Wait()
+	pool.Close()
+	for i := 0; i < 10; i++ {
 		pool.Process(i)
 	}
 	pool.Wait()
-	log.Println("complete")
+	pool.Close()
+	log.Printf("complete, uptime: %s", time.Since(uptime))
+}
+
+func TestQ(t *testing.T) {
+	q := make(chan any)
+	go func() {
+		log.Println("b")
+		log.Println(len(q))
+
+		<-q
+	}()
+
+	log.Println("a")
+	q <- 1
+
+	close(q)
 }
